+++ src/main/java/com/ttubeog/domain/comment/dto/request/WriteCommentReq.java
@@ -9,8 +9,10 @@ public class WriteCommentReq {
 
     @Schema(description = "댓글 내용")
     private String content;
+
     @Schema(description = "위도")
-    private Float latitude;
+    private Double latitude;
+
     @Schema(description = "경도")
-    private Float longitude;
+    private Double longitude;
 }
